+++ src/project/component/shared/materialInstance.cpp
@@ -55,7 +55,7 @@ void Project::Component::Shared::MaterialInstance::build(Utils::BinaryFile &file
   if(setLighting.resolve(obj)) setMask |= 1 << 3;
   if(setFresnel.resolve(obj)) setMask |= 1 << 4;
 
-  for(int i=0; i<texSlots.size(); ++i) {
+  for(size_t i=0; i<texSlots.size(); ++i) {
     if(texSlots[i].set.value) {
       setMask |= 1 << (texSlots.size() + i);
     }
@@ -73,7 +73,7 @@ void Project::Component::Shared::MaterialInstance::build(Utils::BinaryFile &file
   file.writeRGBA(env.resolve(obj));
   file.writeRGBA(fresnelColor.resolve(obj));
 
-  for(int i=0; i<texSlots.size(); ++i)
+  for(size_t i=0; i<texSlots.size(); ++i)
   {
     if(texSlots[i].set.value) {
       file.write<uint32_t>(0); // runtime pointer
